Build arguments and pass them to the logger
This is an effort to try and reduce code bloat due to the templates by
isolating the duplicated code into a constexpr enabled series of
functions during build time, while keeping the logger code free of
templates.

It gives a pretty good improvement of saving 1kB in the test app.
It also improves runtime slightly.

Before
4730          1056     104    5890    1702 build/targets/format
After
3854          1056     104    5014    1396 build/targets/format

Closes #52

Change-Id: Id5e5f65b28f56c50abc677afdf068c3b5c73b265