Disable C_DEBUGEN
For the Debug Monitor ARM exception to trigger, the C_DEBUGEN bit must
be disabled when the debugger is attached.

This means that we won't be able to halt the core or receive debug
events from the host, but this shouldn't be a problem since so far it is
not possible to concurrently multiplex a gdb server and rtt in postform.

Change-Id: I9087cc917f9d02d25ad28b252b89d06657a2c20d