Fix clang-12 build
This issue is not reproducible with clang-10 or clang-11.

clang-12 invokes the linker with -L${SYSROOT}/lib which triggers an
issue because the arm-none-eabi-gcc sysroot has a multilib structure.

telling the linker to find libraries in that directory is
unfortunately wrong as it doesn't really take into account the
architecture of the platform.

This was causing the linker to link against an ARM library, inserting
blx instructions to transition to ARM from Thumb. Since the Cortex-M3
does not implement ARM, this triggers a fault.

Fixes #111

Change-Id: Icfaca0f2024a8e2883c83a683c98e5de87630506