Add CLI option to disable FW version check
Sometimes this is useful when the user knows that the transport format
is compatible regardless of a version difference in the binary.

In the future it would be nice if this information could be derived from
the semantic versioning, e.g.: FW with older minor should be compatible
with a more recent postform_decoder version and differences in patch
numbers should be ignored.

Currently that solution is not as useful because anything can change
before 1.0.0 as the API is not yet stable.

Change-Id: I8669151ede290bc290574d5bc41dbfdb4aced1a9