Reset bit 0 in BP address
When setting a breakpoint for an ARM processor bit 0 is often used to
indicate whether the code we are jumping to is ARM or thumb. Even though
Cortex-M CPUs only support thumb-2 so far, all jump addresses have bit
0 set for this reason.

When setting a breakpoint probe-rs doesn't seem to like this in its
latest release, so let's fix it.